requirements.txt
pytorch
torchvision
cudatoolkit
tqdm
